+++ src/MySql/internal/New-AzMySqlServerAdministrator.ps1
@@ -45,6 +45,7 @@ INPUTOBJECT <IMySqlIdentity>: Identity Parameter
   [DatabaseName <String>]: The name of the database.
   [FirewallRuleName <String>]: The name of the server firewall rule.
   [Id <String>]: Resource identity path
+  [KeyName <String>]: The name of the server key.
   [LocationName <String>]: The name of the location.
   [ResourceGroupName <String>]: The name of the resource group. The name is case insensitive.
   [SecurityAlertPolicyName <SecurityAlertPolicyName?>]: The name of the security alert policy.
@@ -53,9 +54,9 @@ INPUTOBJECT <IMySqlIdentity>: Identity Parameter
   [VirtualNetworkRuleName <String>]: The name of the virtual network rule.
 
 PROPERTY <IServerAdministratorResource>: Represents a and external administrator to be created.
-  Login <String>: The server administrator login account name.
-  Sid <String>: The server administrator Sid (Secure ID).
-  TenantId <String>: The server Active Directory Administrator tenant id.
+  [Login <String>]: The server administrator login account name.
+  [Sid <String>]: The server administrator Sid (Secure ID).
+  [TenantId <String>]: The server Active Directory Administrator tenant id.
 .Link
 https://docs.microsoft.com/en-us/powershell/module/az.mysql/new-azmysqlserveradministrator
 #>