+++ src/Synapse/Synapse/Models/PSDataFlowResource.cs
@@ -1,7 +1,4 @@
 ﻿using Azure.Analytics.Synapse.Artifacts.Models;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace Microsoft.Azure.Commands.Synapse.Models
 {
@@ -14,16 +11,11 @@ namespace Microsoft.Azure.Commands.Synapse.Models
                   dataFlowResource?.Etag)
         {
             this.WorkspaceName = workspaceName;
-            this.Properties = new PSDataFlow(dataFlowResource?.Properties);
+            this.Properties = dataFlowResource?.Properties;
         }
 
         public string WorkspaceName { get; set; }
 
-        public PSDataFlow Properties { get; set; }
-
-        public DataFlowResource ToSdkObject()
-        {
-            return new DataFlowResource(this.Properties?.ToSdkObject());
-        }
+        public DataFlow Properties { get; set; }
     }
 }