+++ src/Synapse/Synapse/Models/PSPipelineResource.cs
@@ -1,15 +1,8 @@
 ﻿using Azure.Analytics.Synapse.Artifacts.Models;
-using Microsoft.Azure.Commands.Common.Compute.Version_2018_04.Models;
 using Microsoft.WindowsAzure.Commands.Utilities.Common;
-using Newtonsoft.Json;
-using Newtonsoft.Json.Linq;
 using Newtonsoft.Json.Serialization;
-using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Reflection;
-using System.Text;
-using System.Text.RegularExpressions;
 
 namespace Microsoft.Azure.Commands.Synapse.Models
 {
@@ -24,7 +17,7 @@ namespace Microsoft.Azure.Commands.Synapse.Models
         {
             this.WorkspaceName = workspaceName;
             this.Description = pipelineResource?.Description;
-            this.Activities = pipelineResource?.Activities?.Select(element => new PSActivity(element)).ToList();
+            this.Activities = pipelineResource?.Activities;
             this.Variables = pipelineResource?.Variables?
                 .Select(element => new KeyValuePair<string, PSVariableSpecification>(element.Key, new PSVariableSpecification(element.Value)))
                 .ToDictionary(kvp => kvp.Key, kvp => kvp.Value);
@@ -50,60 +43,22 @@ namespace Microsoft.Azure.Commands.Synapse.Models
 
         public string WorkspaceName { get; set; }
 
-        [JsonProperty(PropertyName = "properties.description")]
         public string Description { get; set; }
 
-        [JsonProperty(PropertyName = "properties.activities")]
-        public IList<PSActivity> Activities { get; set; }
+        public IList<Activity> Activities { get; set; }
 
-        [JsonProperty(PropertyName = "properties.variables")]
         public IDictionary<string, PSVariableSpecification> Variables { get; set; }
 
-        [JsonProperty(PropertyName = "properties.concurrency")]
         public int? Concurrency { get; set; }
 
-        [JsonProperty(PropertyName = "properties.annotations")]
         public IList<object> Annotations { get; set; }
 
-        [JsonProperty(PropertyName = "properties.runDimensions")]
         public IDictionary<string, object> RunDimensions { get; set; }
 
-        [JsonProperty(PropertyName = "properties.folder")]
         public PSPipelineFolder Folder { get; set; }
 
-        [JsonProperty(PropertyName = "properties.parameters")]
         public IDictionary<string, PSParameterSpecification> Parameters { get; set; }
         
-        [JsonProperty(PropertyName = "properties")]
-        [JsonExtensionData]
         public IDictionary<string, object> AdditionalProperties { get; set; }
-
-        public PipelineResource ToSdkObject()
-        {
-            PipelineResource pipeline = new PipelineResource
-            {
-                Description = this.Description,
-                Concurrency = this.Concurrency,
-                Folder = this.Folder?.ToSdkObject()
-            };
-            this.Activities?.ForEach(item => pipeline.Activities.Add(item?.ToSdkObject()));
-            this.Variables?.ForEach(item => pipeline.Variables.Add(item.Key, item.Value?.ToSdkObject()));
-            this.Annotations?.ForEach(item => pipeline.Annotations.Add(item));
-            this.RunDimensions?.ForEach(item => pipeline.RunDimensions.Add(item));
-            this.Parameters?.ForEach(item => pipeline.Parameters.Add(item.Key, item.Value?.ToSdkObject()));
-            List<string> properties = new List<string> { "description", "activities", "variables", "concurrency",
-                "annotations", "runDimensions", "folder", "parameters" };
-            if (this.AdditionalProperties != null)
-            {
-                foreach (var item in this.AdditionalProperties)
-                {
-                    if (!properties.Contains(item.Key))
-                    {
-                        pipeline.Add(item.Key, item.Value);
-                    }
-                }
-            }
-            return pipeline;
-        }
     }
 }