+++ src/Databricks/ChangeLog.md
@@ -19,6 +19,9 @@
 -->
 ## Upcoming Release
 
+## Version 1.0.2
+* Fixed an issue that may cause `New-AzDatabricksVNetPeering` to return before it is fully provisioned (https://github.com/Azure/autorest.powershell/issues/610)
+
 ## Version 1.0.1
 * Fixed a bug that may cause updating databricks workspace without `-EncryptionKeyVersion` to fail.